refactor(escalas): extract EscalaCard, EscalaFiltrosSheet, AdicionarEscalaDialog
